+++ Changelog.md
@@ -56,6 +56,7 @@
 * Remove usage of deprecated `Coordinates#get_column_number`
 * Use instance method `Relationships#parse_file` instead of class method `Relationships.parse_rels`
 * Simplify parsing `DocxParagraphRun#font` for `asciiTheme`
+* Redone parsing of `DocxStructure#comments`
   
 ### Fixes
 * Fix crash on empty coordinates list of chart